+++ lib/code_box.dart
@@ -2,3 +2,5 @@ library code_box;
 
 export 'src/CodeBoxList.dart' show CodeBoxList;
 export 'src/CodeBoxWidgetModifier.dart' show CodeBoxWidgetModifier;
+export 'src/CodeBoxIdentifiable.dart' show Identifiable;
+export 'src/CodeBoxSet.dart' show CodeBoxSet;