+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.1
+again export fix, been a while
+
+## 1.1.0
+Add CodeBox Identifiable and better set usages
+
 ## 1.0.1
 Export fix, show WidgetModifier